--- lib/presentation/chat/widgets/chat_list_appbar.dart
@@ -1,51 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:wisp/core/config/constants/colors.dart';
-
-class ChatListAppbar extends StatelessWidget implements PreferredSizeWidget {
-  const ChatListAppbar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      backgroundColor: WispColors.deepBlue1,
-      title: const Text(
-        '채팅',
-        style: TextStyle(
-          color: Colors.white,
-          fontSize: 20,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      centerTitle: false,
-      actions: [
-        IconButton(
-          onPressed: () {},
-          icon: const Icon(Icons.add),
-          color: Colors.white,
-        )
-      ],
-      bottom: PreferredSize(
-        preferredSize: const Size.fromHeight(48),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-          child: TextField(
-            decoration: InputDecoration(
-              prefixIcon: const Icon(Icons.search, color: Colors.grey),
-              hintText: '채팅방 또는 메시지 검색',
-              filled: true,
-              fillColor: Colors.white,
-              contentPadding: const EdgeInsets.symmetric(horizontal: 12),
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(12),
-                borderSide: BorderSide.none,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  Size get preferredSize => const Size.fromHeight(kToolbarHeight + 48);
-}
